+++ scripts/recipe_robot_lib/tools.py
@@ -385,6 +385,15 @@ def get_user_defaults():
 
 
 def save_user_defaults(prefs):
+    # Clean up non Recipe Robot related keys that were accidentally collected from the
+    # global preferences by prior versions of Recipe Robot.
+    cfprefs_keylist = CFPreferencesCopyKeyList(
+        BUNDLE_ID, kCFPreferencesCurrentUser, kCFPreferencesAnyHost
+    )
+    external_keys = [x for x in cfprefs_keylist if x not in PREFERENCE_KEYS]
+    for ext_key in external_keys:
+        CFPreferencesSetAppValue(ext_key, None, BUNDLE_ID)
+
     # Save latest values for all Recipe Robot keys.
     for key in PREFERENCE_KEYS:
         if prefs.get(key):